+++ src/cpu/x64/lrn/lrn_avx512_nhwc_executor.hpp
@@ -32,11 +32,11 @@ class lrn_avx512_nhwc_executor_fwd_t : public i_lrn_executor_t {
 public:
     lrn_avx512_nhwc_executor_fwd_t(const PD_T *pd)
         : ker_(utils::make_unique<
-                  lrn::jit_avx512_common_lrn_kernel_fwd_nhwc_t<d_type>>(pd->C(),
-                  pd->desc()->prop_kind,
+                  lrn::jit_avx512_common_lrn_kernel_fwd_nhwc_t<d_type>>(
+                  static_cast<unsigned>(pd->C()), pd->desc()->prop_kind,
                   pd->desc()->lrn_alpha / pd->desc()->local_size,
                   pd->desc()->lrn_beta, pd->desc()->lrn_k,
-                  pd->desc()->local_size))
+                  static_cast<int>(pd->desc()->local_size)))
         , N_(pd->MB())
         , C_(pd->C())
         , H_(pd->H())
@@ -88,9 +88,11 @@ class lrn_avx512_nhwc_executor_bwd_t : public i_lrn_executor_t {
 public:
     lrn_avx512_nhwc_executor_bwd_t(const PD_T *pd)
         : ker_ {utils::make_unique<
-                  lrn::jit_avx512_common_lrn_kernel_bwd_nhwc_t<d_type>>(pd->C(),
+                  lrn::jit_avx512_common_lrn_kernel_bwd_nhwc_t<d_type>>(
+                  static_cast<unsigned>(pd->C()),
                   pd->desc()->lrn_alpha / pd->desc()->local_size,
-                  pd->desc()->lrn_beta, pd->desc()->local_size)}
+                  pd->desc()->lrn_beta,
+                  static_cast<int>(pd->desc()->local_size))}
         , N_(pd->MB())
         , C_(pd->C())
         , H_(pd->H())
